Refactor protocol code

+++ contracts/stable_coin_factory/sources/kasa_manager.move
@@ -35,8 +35,9 @@ module stable_coin_factory::kasa_manager {
     use tokens::rusd_stable_coin::{Self, RUSD_STABLE_COIN, RUSDStableCoinStorage};
     use library::kasa::{get_minimum_collateral_ratio, calculate_nominal_collateral_ratio, get_minimum_net_debt};
     use library::math::{min, mul_div, scalar};
-    use oracles::oracle::{get_sui_price};
-    use SupraOracle::SupraSValueFeed::{Self, OracleHolder};
+
+    use oracles::oracle::get_sui_price;
+    use SupraOracle::SupraSValueFeed::OracleHolder;
   
     // use library::utils::logger;
 
@@ -381,7 +382,7 @@ module stable_coin_factory::kasa_manager {
     ) {
         // TODO: Take a look at the stake logic - what does it do?
 
-        let collateral_price = (get_sui_price(oracle_holder) as u64); // FIXME: Change this to the actual price
+        let collateral_price = get_sui_price(oracle_holder);
         //let collateral_price = 1600_000000000;
 
         let stability_pool_stake_amount = stability_pool::get_total_stake_amount(sp_storage);
@@ -465,8 +466,7 @@ module stable_coin_factory::kasa_manager {
         // TODO: Require max valid fee percentage -> IMPLEMENT FEES
         // TODO: Disable this method for 14 days after release
 
-        let collateral_price = (get_sui_price(oracle_holder) as u64); // FIXME: Change this to the actual price
-        //let collateral_price = 1600_000000000;
+        let collateral_price = get_sui_price(oracle_holder);
 
         let user_stable_coin_balance = rusd_stable_coin::get_balance(rsc_storage, account_address);
